requirements.txt
flask==2.3.3
paramiko==3.2.0
python-dotenv==1.0.0